pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic_file_organizer"
version = "1.0.0"
description = "AI agent environment for real-world file categorization."
authors = [
    {name = "Kanak Megha"}
]
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "openai",
    "requests",
    "openenv-core"
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
where = ["."]